requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==5.2.9
django-browser-reload==1.21.0
django-jazzmin==3.0.1
django-phonenumber-field==8.4.0
django-tailwind==4.4.1
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
phonenumbers==9.0.20
pillow==12.0.0
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
